Fix review findings in external input hardening
- Uri.pathSegments/queryParameters percent-decode lazily and throw
  FormatException on malformed escapes, which Uri.tryParse accepts. Both
  PaymentIntent.tryParseUrl and DeepLinkService.handleLink read them
  unguarded, so a /pay?to=%c3%28 link threw out of the app_links stream
  listener and out of the barcode callback. Route every read through a
  fail-closed accessor.
- DecimalInputFilter had no length cap, so a million-digit paste entered
  the TextField and was only rejected after text layout. Enforce
  LocaleNumberConfig.maxInputLength in the filter.
- isAcceptableUrPart treated a sequence header with 7+ digits as a
  header-less single-part frame and accepted it, so the fragment-count cap
  only fired on honest input. Anchor the header to the frame head, parse it
  with int.tryParse, and reject an index outside its own declared count the
  way the decoder does.
- encodeUrForQr could emit more frames than the scanners accept; fail at
  encode time instead.
- Hitting the accumulation cap dropped every further frame in silence.
  Log it, reset the accumulator, and surface an error; make the frame
  budget an AnimatedQrScanner parameter so the widget stays format-agnostic.
- Log the dropped code in QrScannerPage instead of returning silently.
- parseDecimal: check the length cap before normalizing, and drop the
  unreachable Decimal.tryParse null branch.
- parseWireAmount: delegate scaling to parseAmount instead of duplicating it.
- Cache the two Rust caps instead of an FFI round-trip per camera frame.
- Drop the usize-to-usize casts that fail cargo clippy -D warnings.

+++ quantus_sdk/rust/src/api/ur.rs
@@ -23,8 +23,7 @@ pub fn max_ur_part_chars() -> u32 {
 }
 
 fn within_limits(ur_parts: &[String]) -> bool {
-    ur_parts.len() <= MAX_UR_PARTS as usize
-        && ur_parts.iter().all(|p| p.len() <= MAX_UR_PART_CHARS as usize)
+    ur_parts.len() <= MAX_UR_PARTS && ur_parts.iter().all(|p| p.len() <= MAX_UR_PART_CHARS)
 }
 
 // Note decode_ur takes the list of QR Codes in any order and assembles them correctly.
